refactor compare_md5 and move functions to common.py
